Added more test cases
